// backend/src/generated/prisma/models/Invoice.ts
export type InvoiceModel = runtime.Types.Result.DefaultSelection<Prisma.$InvoicePayload>
// ...
export type $InvoicePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Invoice"
  objects: {
    order: Prisma.$OrderPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    orderId: string
    number: string
    fileUrl: string | null
    issuedAt: Date
    paidAt: Date | null
  }, ExtArgs["result"]["invoice"]>
  composites: {}
}